Cache icon data
On Unix there is no problem in saving the pixmap and opening the file immediately after
to read its content. On Windows, however, the file is always half-written so we get
corrupted data. Cache the icon binary representation and use it in get_image instead
of opening and reading the file directly.